Added JSON scrapping
